add dataloader for pretraining
